Conversion of this plugin to Joomla3
The module seems to be working good, but the component still is not
creating the meeting rooms. If the meeting is created manually in MySQL
everything seems to work fine.

The component is working on my development environment up to the point
that you can change and add salt and URL of bbb server.